[EOS-18240] adds test api helpers and examples (#1143)
* adds test api helpers and examples


* makes codacy happy
