+++ test/debug/test_debug.py
@@ -22,15 +22,11 @@ import logging
 logger = logging.getLogger(__name__)
 
 
-# FIXME debug
 @pytest.mark.debug
-#@pytest.mark.skip
 def test_debug_fail(any_level, any_topic):
     pytest.fail()
 
 
-# FIXME debug
 @pytest.mark.debug
-#@pytest.mark.skip
 def test_debug_pass(any_level, any_topic):
     pass
