+++ test/testapi/__init__.py
@@ -15,29 +15,25 @@
 # please email opensource@seagate.com or cortx-questions@seagate.com.
 #
 
-import pytest
-from typing import List
-from enum import Enum
+from .integration import (
+    root_passwd,
+    hosts_num,
+    hosts_num_module,
+    setup_hosts,
+    HostMeta
+)
 
+fixtures = [
+    root_passwd,
+    hosts_num,
+    hosts_num_module,
+    setup_hosts,
+]
 
-from test.conftest import HostMeta
-
-from test.testapi import defs
-
-
-class RunT(Enum):
-    """Modes of setup run"""
-
-    REMOTE_CLI = 'remote_cli'       # logic is run on the host system via CLI
-    REMOTE_API = 'remote_api'       # via API
-    ONTARGET_CLI = 'ontarget_cli'   # logic is run on a target system via CLI
-    # ONTARGET_API = 'ontarget_api' # via API
-
-ScaleFactorT = defs.ScaleFactorT
-
-
-class SourceT(Enum):
-    """Types of sources for setup"""
-
-    LOCAL = 'local'
-    ISO = 'iso'
+__all__ = [
+    'root_passwd',
+    'hosts_num',
+    'hosts_num_module',
+    'setup_hosts',
+    'HostMeta'
+]
